sidecar-python/requirements-dev.txt
grpcio-tools>=1.60.0
pytest>=7.0
python-pptx>=1.0.0
xlwt>=1.3.0
